+++ gunicorn/workers/ggevent.py
@@ -77,7 +77,7 @@ class GeventWorker(AsyncWorker):
         try:
             # Try to stop connections until timeout
             self.notify()
-            server.stop(timeout=self.timeout)
+            server.stop(timeout=self.cfg.graceful_timeout)
         except:
             pass
 
@@ -87,11 +87,12 @@ class GeventWorker(AsyncWorker):
         except gevent.GreenletExit:
             pass
 
-    if hasattr(gevent.core, 'dns_shutdown'):
+    if gevent.version_info[0] == 0:
 
         def init_process(self):
             #gevent 0.13 and older doesn't reinitialize dns for us after forking
             #here's the workaround
+            import gevent.core
             gevent.core.dns_shutdown(fail_requests=1)
             gevent.core.dns_init()
             super(GeventWorker, self).init_process()