+++ tests/002-test-invalid-requests.py
@@ -8,11 +8,21 @@ import treq
 
 import glob
 import os
+from nose.tools import raises
+
 dirname = os.path.dirname(__file__)
 reqdir = os.path.join(dirname, "requests", "invalid")
 
+
 def test_http_parser():
     for fname in glob.glob(os.path.join(reqdir, "*.http")):
-        expect = treq.load_py(os.path.splitext(fname)[0] + ".py")
-        req = treq.badrequest(fname, expect)
-        yield (req.check,)
+        env = treq.load_py(os.path.splitext(fname)[0] + ".py")
+        expect = env['request']
+        cfg = env['cfg']
+        req = treq.badrequest(fname)
+
+        @raises(expect)
+        def check(fname):
+            return req.check(cfg)
+
+        yield check, fname # fname is pass so that we know which test failed