+++ tests/requests/invalid/011.py
@@ -1,0 +1,6 @@
+from gunicorn.config import Config
+from gunicorn.http.errors import LimitRequestHeaders
+
+request = LimitRequestHeaders
+cfg = Config()
+cfg.set('limit_request_fields', 2)